LPS-177038 move data items to hooks/data
