+++ modules/apps/site-initializer/site-initializer-testray/site-initializer-testray/extra/remote-app/src/hooks/data/useBuildHistory.ts
@@ -12,9 +12,9 @@
  * details.
  */
 
-import {TestrayBuild} from '../services/rest';
-import {DATA_COLORS, Statuses} from '../util/constants';
-import {getRandomMaximumValue} from '../util/mock';
+import {TestrayBuild} from '../../services/rest';
+import {DATA_COLORS, Statuses} from '../../util/constants';
+import {getRandomMaximumValue} from '../../util/mock';
 
 const useBuildHistory = () => {
 	return {
