+++ modules/apps/site-initializer/site-initializer-testray/site-initializer-testray/extra/remote-app/src/hooks/data/useTableChartData.ts
@@ -12,7 +12,7 @@
  * details.
  */
 
-import {translate} from '../i18n';
+import {translate} from '../../i18n';
 
 type TableChartDataProps = {
 	colors: string[][];
