+++ modules/apps/site-initializer/site-initializer-testray/site-initializer-testray/extra/remote-app/src/hooks/data/useTotalTestCases.ts
@@ -12,8 +12,8 @@
  * details.
  */
 
-import {Statuses, chartColors} from '../util/constants';
-import {getRandomMaximumValue} from '../util/mock';
+import {Statuses, chartColors} from '../../util/constants';
+import {getRandomMaximumValue} from '../../util/mock';
 
 const useTotalTestCases = () => {
 	const donutColumns = [
